Add files via upload
This file contains c++ program code for a simple game known as chasing the rabbit. The program allows user to                                                       choose their respective rows and columns and begin the game to chase the rabbit. Once, the game starts the user                                                                                                                                              is allowed to move in only one direction & only one row or column at a time. After, every move the rabbit makes                                                                                                                  its moves and the direction it moved is notified to the user. After, every move the program checks if the user                                                                                                                                                  wins the game or not, and notifies accordingly. In the end, the user is asked if they want to play again.